@@ -1,112 +1,129 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
+import { APP_INITIALIZER, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, APP_INITIALIZER } from '@angular/core';
-import { LocationStrategy, HashLocationStrategy, PathLocationStrategy } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { LayoutModule } from '@angular/cdk/layout';
+import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 
-import { OidcConfigService, AuthModule } from 'angular-auth-oidc-client';
-import { PerfectScrollbarModule } from 'ngx-perfect-scrollbar';
-import { PerfectScrollbarConfigInterface } from 'ngx-perfect-scrollbar';
-// RECOMMENDED
-import { CarouselModule } from 'ngx-bootstrap/carousel';
+import { AmplifyAuthenticatorModule } from '@aws-amplify/ui-angular';
 
-const DEFAULT_PERFECT_SCROLLBAR_CONFIG: PerfectScrollbarConfigInterface = {
-  suppressScrollX: true
-};
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatNativeDateModule } from '@angular/material/core';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { UnauthorizedComponent } from './views/error/unauthorized.component';
+
+import { map, shareReplay, switchMap } from 'rxjs/operators';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
-
-// Import containers
-import { DefaultLayoutComponent } from './default-layout';
-
-import { P404Component } from './views/error/404.component';
-import { P500Component } from './views/error/500.component';
-import { LogoffComponent } from './views/logoff/logoff.component';
-
-
-const APP_CONTAINERS = [
-  DefaultLayoutComponent
-];
-
-import {
-  AppAsideModule,
-  AppBreadcrumbModule,
-  AppHeaderModule,
-  AppFooterModule,
-  AppSidebarModule,
-} from '@coreui/angular';
-
-// Import routing module
-import { AppRoutingModule } from './app.routing';
-
-// Import 3rd party components
-import { AlertModule } from 'ngx-bootstrap/alert';
-import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
-import { TabsModule } from 'ngx-bootstrap/tabs';
-import { ChartsModule } from 'ng2-charts';
-import { HttpClient } from '@angular/common/http';
-import { configureAuth } from './views/auth/auth-config';
-import { UnauthorizedComponent } from './views/error/unauthorized.component';
+import { AppRoutingModule } from './app-routing.module';
+import { ConfigAssetLoaderService } from 'config-asset-loader';
+import { NavComponent } from './nav/nav.component';
+import { Amplify } from 'aws-amplify';
+import { AuthComponent } from './views/auth/auth.component';
 import { httpInterceptorProviders } from './interceptors';
-import { CollapseModule } from 'ngx-bootstrap/collapse';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { AuthConfigurationService } from './views/auth/auth-configuration.service';
-
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatDialogModule } from '@angular/material/dialog';
+import {
+  MatFormFieldModule,
+  MAT_FORM_FIELD_DEFAULT_OPTIONS,
+} from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatRadioModule } from '@angular/material/radio';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTableModule } from '@angular/material/table';
 
 @NgModule({
-  imports: [
-    AlertModule.forRoot(),
-    AppAsideModule,
-    AppBreadcrumbModule.forRoot(),
-    AppFooterModule,
-    AppHeaderModule,
-    AppRoutingModule,
-    AppSidebarModule,
-    AuthModule.forRoot(),
-    BrowserAnimationsModule,
-    BrowserModule,
-    BsDropdownModule.forRoot(),
-    CarouselModule.forRoot(),
-    ChartsModule,
-    CollapseModule.forRoot(),
-    FormsModule,
-    PerfectScrollbarModule,
-    ReactiveFormsModule,
-    TabsModule.forRoot(),
-  ],
   declarations: [
-    ...APP_CONTAINERS,
     AppComponent,
-    LogoffComponent,
-    P404Component,
-    P500Component,
+    NavComponent,
+    AuthComponent,
     UnauthorizedComponent,
   ],
+  imports: [
+    AmplifyAuthenticatorModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    BrowserModule,
+    HttpClientModule,
+    LayoutModule,
+    MatButtonModule,
+    MatCardModule,
+    MatGridListModule,
+    MatIconModule,
+    MatListModule,
+    MatMenuModule,
+    MatNativeDateModule,
+    MatProgressSpinnerModule,
+    MatSidenavModule,
+    MatToolbarModule,
+    MatSnackBarModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatButtonModule,
+    MatCardModule,
+    MatIconModule,
+    MatDatepickerModule,
+    MatDialogModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatRadioModule,
+    MatSelectModule,
+    MatTableModule,
+  ],
   providers: [
-    HttpClient,
+    ConfigAssetLoaderService,
+    HttpClientModule,
     {
       provide: LocationStrategy,
       useClass: HashLocationStrategy,
     },
-    OidcConfigService,
+    // {
+    //   provide: APP_INITIALIZER,
+    //   useFactory: InitAuthSettings,
+    //   deps: [HttpClient],
+    //   multi: true,
+    // },
     httpInterceptorProviders,
-    AuthConfigurationService,
   ],
-  bootstrap: [ AppComponent ]
+  bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
+
+export function InitAuthSettings(http: HttpClient) {
+  console.log('In INITAUTHSETTINGS');
+  return () => {
+    return http.get<Configuration>('./assets/config/config.json').pipe(
+      switchMap((config) => {
+        console.log('CONFIG: ', JSON.stringify(config));
+        const apiUrl = `${config.apiUrl}/api/tenants/auth-info`;
+        return http.get<AuthInfo>(apiUrl);
+      }),
+      map((authInfo) => {
+        console.log(authInfo);
+        Amplify.configure(authInfo);
+      }),
+      shareReplay(1)
+    );
+  };
+}
+
+interface Configuration {
+  apiUrl: string;
+  stage: string;
+}
+
+export interface AuthInfo {
+  aws_project_region: string;
+  aws_cognito_region: string;
+  aws_user_pools_id: string;
+  aws_user_pools_web_client_id: string;
+}
