@@ -1,5 +1,5 @@
 export const environment = {
-  production: true,
-  regApiGatewayUrl: 'https://3l6yegpnyc.execute-api.us-west-2.amazonaws.com/prod/',
-  domain: ''
+  production: false,
+  regApiGatewayUrl:
+    'https://wchv3gaofb.execute-api.us-west-2.amazonaws.com/prod',
 };
